scraper_function/requirements.txt
google-cloud-pubsub==2.24.2
google-cloud-storage==2.19.0
journ4list==0.3.0
google-api-core[grpc]==2.24.2
